python/requirements.txt
psycopg2
python-dotenv
requests==2.32.3
